Add console sample app offering C++ interface
In case of native apps, list of libraries and samples is limited to
Android, iOS & Windows (this repository):

	https://developers.google.com/identity/protocols/oauth2/native-app#libraries

Google APIs Client Library for C++ exists but it is deprecated as of
May 30th 2019. To be able to develop new projects requiring C++ without
usage of deprecated library and without the need of maintaining your
own custom solution, you can rely on available client libraries listed
here:

	https://developers.google.com/identity/protocols/oauth2#libraries

and prepare just a thin wrapper. Core logic remains mostly unchanged
in comparison to pure C# based console (CLI) sample application. To
use it in standard C++, thin layer of CLR is added in-between.